% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tobitpolyvar.R
\name{tobitpolyvar}
\alias{tobitpolyvar}
\title{Perform tobit regression assuming polynomial-based heteroscedasticity}
\usage{
tobitpolyvar(formula, data, left = -1, start.beta = NULL, start.gamma)
}
\arguments{
\item{formula}{a formula specifying the relationship between the tobit response and regressors}

\item{data}{a data.frame containing the variables specified in formula}

\item{left}{a number specifying where left-censoring occurred}

\item{start.beta}{an initial guess for beta}

\item{start.gamma}{an initial guess for gamma. This also implicity specifies the degree of the polynomial, i.e. length(gamma) - 1}
}
\value{
a list containing the following parameter estimates (from maximum likelihood):\cr
\item{beta}{the regression coefficients}
\item{gamma}{the polynomial coefficients}
}
\description{
This function gets maximum-likelihood estimates of beta and gamma assuming the data follow a
tobit model with polynomial-based heteroscedasticity. NOTE: This is only for use in the EQ5D5L
project at SickKids. There is NO guarantee it will work for a general use-case.
}
